+++ app/mypage/MyPageClient.tsx
@@ -303,67 +303,130 @@ export default function MyPageClient({ user, profile, posts, comments, interview
 
           {activeTab === 'interviews' && (
             <div className="max-w-4xl">
-              <h1 className="text-2xl font-bold text-gray-900 dark:text-white mb-8">면접 후기</h1>
+              <h1 className="text-2xl font-bold text-gray-900 dark:text-white mb-8">내 면접 후기</h1>
               
               {interviews && interviews.length > 0 ? (
-                <div className="space-y-6">
+                <div className="grid gap-6">
                   {interviews.map((interview: any) => (
                     <Link key={interview.id} href={`/interview/${interview.id}`}>
-                      <div className="bg-white dark:bg-gray-800 rounded-xl shadow-sm border border-gray-200 dark:border-gray-700 p-6 hover:shadow-md transition-shadow cursor-pointer">
+                      <div className="bg-white dark:bg-gray-800 rounded-2xl shadow-sm border border-gray-200 dark:border-gray-700 p-6 hover:shadow-lg hover:border-blue-300 dark:hover:border-blue-600 transition-all duration-200 cursor-pointer group">
                         <div className="flex items-start justify-between mb-4">
                           <div className="flex-1">
-                            <h3 className="text-lg font-semibold text-gray-900 dark:text-white mb-2">
-                              {interview.company_name} - {interview.position}
-                            </h3>
-                            <div className="flex items-center gap-4 text-sm text-gray-500 dark:text-gray-400">
+                            <div className="flex items-center gap-3 mb-3">
+                              <h3 className="text-xl font-bold text-gray-900 dark:text-white group-hover:text-blue-600 dark:group-hover:text-blue-400 transition-colors">
+                                {interview.company_name}
+                              </h3>
+                              <span className="text-gray-400 dark:text-gray-500">•</span>
+                              <span className="text-lg font-medium text-gray-700 dark:text-gray-300">
+                                {interview.position}
+                              </span>
+                            </div>
+                            <div className="flex flex-wrap items-center gap-4 text-sm text-gray-600 dark:text-gray-400">
                               {interview.interview_date && (
-                                <span>면접일: {new Date(interview.interview_date).toLocaleDateString('ko-KR')}</span>
+                                <div className="flex items-center gap-1">
+                                  <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z" />
+                                  </svg>
+                                  <span>면접일: {new Date(interview.interview_date).toLocaleDateString('ko-KR')}</span>
+                                </div>
                               )}
                               {interview.interview_type && (
-                                <span>유형: {interview.interview_type}</span>
+                                <div className="flex items-center gap-1">
+                                  <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9.663 17h4.673M12 3v1m6.364 1.636l-.707.707M21 12h-1M4 12H3m3.343-5.657l-.707-.707m2.828 9.9a5 5 0 117.072 0l-.548.547A3.374 3.374 0 0014 18.469V19a2 2 0 11-4 0v-.531c0-.895-.356-1.754-.988-2.386l-.548-.547z" />
+                                  </svg>
+                                  <span>유형: {
+                                    interview.interview_type === '화상면접' ? '화상면접' :
+                                    interview.interview_type === '대면면접' ? '대면면접' :
+                                    interview.interview_type === '전화면접' ? '전화면접' :
+                                    interview.interview_type === '기타' ? '기타' :
+                                    interview.interview_type === 'video' ? '화상면접' :
+                                    interview.interview_type === 'in_person' ? '대면면접' :
+                                    interview.interview_type === 'phone' ? '전화면접' :
+                                    interview.interview_type === 'technical' ? '기술면접' :
+                                    interview.interview_type === 'behavioral' ? '인성면접' :
+                                    interview.interview_type === 'case_study' ? '케이스 스터디' :
+                                    interview.interview_type === 'presentation' ? '발표면접' : 
+                                    interview.interview_type === 'group' ? '그룹면접' :
+                                    interview.interview_type === 'other' ? '기타' :
+                                    interview.interview_type || '기타'
+                                  }</span>
+                                </div>
                               )}
                               {interview.difficulty_level && (
-                                <span>난이도: {interview.difficulty_level}</span>
+                                <div className="flex items-center gap-1">
+                                  <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
+                                  </svg>
+                                  <span>난이도: {
+                                    interview.difficulty_level === 'easy' ? '쉬움' :
+                                    interview.difficulty_level === 'medium' ? '보통' :
+                                    interview.difficulty_level === 'hard' ? '어려움' :
+                                    interview.difficulty_level === 'very_hard' ? '매우 어려움' : interview.difficulty_level
+                                  }</span>
+                                </div>
                               )}
                             </div>
                           </div>
-                          <div className="flex items-center gap-2">
+                          <div className="flex flex-col items-end gap-3">
                             {interview.overall_rating && (
-                              <div className="flex items-center gap-1">
-                                <span className="text-yellow-400">⭐</span>
-                                <span className="text-sm font-medium text-gray-700 dark:text-gray-300">
+                              <div className="flex items-center gap-1 bg-yellow-50 dark:bg-yellow-900/20 px-3 py-1 rounded-full">
+                                <span className="text-yellow-500">⭐</span>
+                                <span className="text-sm font-semibold text-yellow-700 dark:text-yellow-400">
                                   {interview.overall_rating}/5
                                 </span>
                               </div>
                             )}
-                            {/* 면접 타입 표시 */}
                             {interview.interview_type && (
-                              <span className="inline-flex items-center px-2 py-1 rounded-full text-xs font-medium bg-blue-100 text-blue-800 dark:bg-blue-900 dark:text-blue-200">
-                                {interview.interview_type === 'technical' ? '기술면접' :
+                              <span className="inline-flex items-center px-3 py-1 rounded-full text-xs font-medium bg-blue-50 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300 border border-blue-200 dark:border-blue-700">
+                                {interview.interview_type === '화상면접' ? '화상면접' :
+                                 interview.interview_type === '대면면접' ? '대면면접' :
+                                 interview.interview_type === '전화면접' ? '전화면접' :
+                                 interview.interview_type === '기타' ? '기타' :
+                                 interview.interview_type === 'video' ? '화상면접' :
+                                 interview.interview_type === 'in_person' ? '대면면접' :
+                                 interview.interview_type === 'phone' ? '전화면접' :
+                                 interview.interview_type === 'technical' ? '기술면접' :
                                  interview.interview_type === 'behavioral' ? '인성면접' :
                                  interview.interview_type === 'case_study' ? '케이스 스터디' :
-                                 interview.interview_type === 'presentation' ? '발표면접' : '기타'}
+                                 interview.interview_type === 'presentation' ? '발표면접' :
+                                 interview.interview_type === 'group' ? '그룹면접' :
+                                 interview.interview_type === 'other' ? '기타' :
+                                 interview.interview_type || '기타'}
                               </span>
                             )}
                           </div>
                         </div>
                         
-                        {interview.result && (
-                          <div className="mb-3">
-                            <span className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-                              interview.result === '합격' 
-                                ? 'bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-200'
-                                : interview.result === '불합격'
-                                ? 'bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-200'
-                                : 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900 dark:text-yellow-200'
-                            }`}>
-                              {interview.result}
-                            </span>
+                        <div className="flex items-center justify-between pt-4 border-t border-gray-100 dark:border-gray-700">
+                          <div className="flex items-center gap-3">
+                            {interview.result && (
+                              <span className={`inline-flex items-center px-3 py-1 rounded-full text-xs font-semibold ${
+                                interview.result === '합격' || interview.result === 'pass' || interview.result === 'passed'
+                                  ? 'bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-300 border border-green-200 dark:border-green-700'
+                                  : interview.result === '불합격' || interview.result === 'fail' || interview.result === 'failed'
+                                  ? 'bg-red-100 text-red-800 dark:bg-red-900/30 dark:text-red-300 border border-red-200 dark:border-red-700'
+                                  : 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900/30 dark:text-yellow-300 border border-yellow-200 dark:border-yellow-700'
+                              }`}>
+                                {interview.result === 'pass' || interview.result === 'passed' ? '합격' :
+                                 interview.result === 'fail' || interview.result === 'failed' ? '불합격' :
+                                 interview.result === 'pending' ? '대기중' :
+                                 interview.result === 'in_progress' ? '진행중' : interview.result}
+                              </span>
+                            )}
+                            <div className="flex items-center gap-1 text-sm text-gray-500 dark:text-gray-400">
+                              <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
+                              </svg>
+                              <span>작성일: {new Date(interview.created_at).toLocaleDateString('ko-KR')}</span>
+                            </div>
                           </div>
-                        )}
-                        
-                        <div className="text-sm text-gray-500 dark:text-gray-400">
-                          작성일: {new Date(interview.created_at).toLocaleDateString('ko-KR')}
+                          <div className="flex items-center text-blue-600 dark:text-blue-400 group-hover:text-blue-700 dark:group-hover:text-blue-300">
+                            <span className="text-sm font-medium mr-1">자세히 보기</span>
+                            <svg className="w-4 h-4 transform group-hover:translate-x-1 transition-transform" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
+                            </svg>
+                          </div>
                         </div>
                       </div>
                     </Link>
@@ -372,8 +435,8 @@ export default function MyPageClient({ user, profile, posts, comments, interview
               ) : (
                 <div className="text-center py-20">
                   <div className="max-w-md mx-auto">
-                    <div className="w-24 h-24 bg-gradient-to-br from-gray-100 to-gray-200 dark:from-gray-800 dark:to-gray-700 rounded-2xl flex items-center justify-center mx-auto mb-6">
-                      <svg className="w-12 h-12 text-gray-400 dark:text-gray-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <div className="w-24 h-24 bg-gradient-to-br from-blue-50 to-blue-100 dark:from-blue-900/20 dark:to-blue-800/20 rounded-3xl flex items-center justify-center mx-auto mb-6 border border-blue-200 dark:border-blue-700">
+                      <svg className="w-12 h-12 text-blue-500 dark:text-blue-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
                       </svg>
                     </div>
@@ -381,12 +444,16 @@ export default function MyPageClient({ user, profile, posts, comments, interview
                       아직 작성한 면접 후기가 없습니다
                     </h3>
                     <p className="text-gray-600 dark:text-gray-400 mb-8 leading-relaxed">
-                      첫 번째 면접 후기를 작성하고 AI 분석을 받아보세요!
+                      첫 번째 면접 후기를 작성하고 AI 분석을 받아보세요!<br/>
+                      면접 경험을 기록하고 다른 구직자들과 공유할 수 있습니다.
                     </p>
                     <Link
                       href="/interview/analyze"
-                      className="inline-flex items-center px-6 py-3 bg-blue-600 hover:bg-blue-700 text-white font-medium rounded-lg transition-colors"
+                      className="inline-flex items-center px-8 py-4 bg-gradient-to-r from-blue-600 to-blue-700 hover:from-blue-700 hover:to-blue-800 text-white font-semibold rounded-xl transition-all duration-200 shadow-lg hover:shadow-xl transform hover:-translate-y-0.5"
                     >
+                      <svg className="w-5 h-5 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 4v16m8-8H4" />
+                      </svg>
                       면접 후기 작성하기
                     </Link>
                   </div>
